refactor: 패키지 구조 정리 및 ui 구조 개선
