+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:usesCleartextTraffic="true">
 
         <activity
-            android:name=".SplashActivity"
+            android:name=".UI.splash.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -26,12 +26,12 @@
         </activity>
 
         <activity
-            android:name=".LoginActivity"
+            android:name=".UI.login.LoginActivity"
             android:exported="true">
         </activity>
 
         <activity
-            android:name=".MainActivity"
+            android:name=".UI.main.MainActivity"
             android:exported="false">
         </activity>
     </application>
