--- app/src/main/java/com/example/daybreak/AuthRepository.kt
@@ -1,9 +1,0 @@
-package com.example.daybreak
-
-import retrofit2.Response
-
-class AuthRepository(private val authService: AuthRetrofitInterface) {
-    suspend fun login(loginRequest: Login): retrofit2.Response<LoginResponse> {
-        return authService.login(loginRequest)
-    }
-}
