+++ app/src/main/java/com/example/daybreak/Data/WeeklyGoalData.kt
@@ -1,4 +1,4 @@
-package com.example.daybreak
+package com.example.daybreak.Data
 
 
 data class WeeklyGoalData(
