--- app/src/main/java/com/example/daybreak/GoalFlowStep2Fragment.kt
@@ -1,4 +1,0 @@
-package com.example.daybreak
-
-class GoalFlowStep2Fragment {
-}
