+++ app/src/main/java/com/example/daybreak/UI/adapter/GoalThisWeekAdapter.kt
@@ -1,3 +1,5 @@
+package com.example.daybreak.UI.adapter
+
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
