+++ app/src/main/java/com/example/daybreak/UI/fragment/FuturemeFragment.kt
@@ -1,10 +1,11 @@
-package com.example.daybreak
+package com.example.daybreak.UI.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.example.daybreak.UI.main.MainActivity
 import com.example.daybreak.databinding.FragmentFuturemeBinding
 
 class FuturemeFragment : Fragment() {
