+++ app/src/main/java/com/example/daybreak/UI/fragment/GoalFlowStep2Fragment.kt
@@ -1,0 +1,4 @@
+package com.example.daybreak.UI.fragment
+
+class GoalFlowStep2Fragment {
+}
