+++ app/src/main/java/com/example/daybreak/UI/splash/SplashActivity.kt
@@ -1,4 +1,4 @@
-package com.example.daybreak
+package com.example.daybreak.UI.splash
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,6 +6,8 @@ import android.os.Handler
 import android.os.Looper
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
+import com.example.daybreak.R
+import com.example.daybreak.UI.login.LoginActivity
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?){
