+++ app/src/main/res/drawable/bg_rounded_gray900_16dp.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <!-- 배경색 -->
+    <solid android:color="@color/Gray_900" />
+
+    <!-- 모서리 둥글기 -->
+    <corners android:radius="16dp" />
+
+</shape>
