+++ app/src/main/res/layout/fragment_goal_flow_step2.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context="com.example.daybreak.GoalFlowStep2Fragment">
+    tools:context="com.example.daybreak.UI.fragment.GoalFlowStep2Fragment">
 
     <!-- 상단 네비게이션 바 -->
     <include
